+++ src/config/events.ts
@@ -3,4 +3,5 @@ export const PICKER_BLUR = 'PICKER_BLUR'
 export const URL_RECEIVED = 'URL_RECEIVED'
 export const FAV_SET = 'FAV_SET'
 export const ACTIVITY_RUN = 'ACTIVITY_RUN'
+export const MOUSE_EVENTS_IGNORE = 'MOUSE_EVENTS_IGNORE'
 export const WINDOW_HIDE = 'WINDOW_HIDE'
