--- src/picker/styledComponents/CopyButton.ts
@@ -1,11 +1,0 @@
-import styled from 'styled-components'
-
-export const CopyButton = styled.button`
-  background-color: transparent;
-  border: none;
-  color: white;
-
-  &:focus {
-    outline: none;
-  }
-`
