--- src/picker/styledComponents/Key.ts
@@ -1,12 +1,0 @@
-import styled from 'styled-components'
-
-export const Key = styled.div`
-  position: absolute;
-  bottom: 0;
-  right: 0;
-  left: 0;
-  text-align: center;
-  color: white;
-  font-size: 18px;
-  font-weight: 400;
-`
