--- src/picker/styledComponents/Url.ts
@@ -1,15 +1,0 @@
-import styled from 'styled-components'
-
-export const Url = styled.div`
-  color: #fafafa;
-  font-size: 20px;
-  line-height: 1.5;
-  margin-top: 30px;
-  padding: 20px 30px;
-  text-align: center;
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  width: 100%;
-  font-family: sans-serif;
-`
